requirements.txt
Flask==3.0.0
azure-identity==1.15.0
azure-mgmt-network==25.0.0
azure-mgmt-subscription==3.1.1
python-dotenv==1.0.0